requirements.txt
fastapi
uvicorn
websockets
passlib[bcrypt,argon2]
python-jose[cryptography]
python-multipart
aiofiles
sqlalchemy
geopybash